Added missing documentation.
Improved API module structure.
