+++ api/swim_api/src/downlink/mod.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use std::{sync::Arc, time::Duration};
+use std::sync::Arc;
 
 use futures::future::BoxFuture;
 use swim_model::path::AbsolutePath;
@@ -20,20 +20,38 @@ use swim_utilities::io::byte_channel::{ByteReader, ByteWriter};
 
 use crate::error::DownlinkTaskError;
 
+/// Indicates the kind of the downlink.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum DownlinkKind {
+    /// Accepts single values and maintains an internal state.
     Value,
+    // Accepts single values but has no state.
     Event,
+    /// Accepts key-value pairs and maintains a state as a map.
     Map,
 }
 
+/// General downlink configuration parameters.
 pub struct DownlinkConfig {
-    _flush_timeout: Duration,
+    //TODO To be filled in later.
 }
 
+/// Trait to define a consumer to a downlink. Instaces of this will be passed to the runtime
+/// to be executed. User code should not generally need to implement this directly. It is
+/// necessary for this trait to be object safe and any changes to it should take that into
+/// account.
 pub trait Downlink {
+    /// The kind of the downlink, used by the runtime to configure how the downlink is managed.
     fn kind(&self) -> DownlinkKind;
 
+    /// Create a task that will manage the state of the downlink and service anything that is
+    /// observing it/ pushing data to it.
+    ///
+    /// #Arguments
+    /// * `path` - The path to the lane to which the downlink should be attached.
+    /// * `config` - Configuration parameters for the downlink task.
+    /// * `input` - Byte channel on which updates will be received from the runtime.
+    /// * `output` - Byte channel on which command will be sent to the runtime.
     fn run(
         &self,
         path: AbsolutePath,
