+++ api/swim_api/src/protocol/downlink/tests.rs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use crate::protocol::{
+use crate::protocol::downlink::{
     DownlinkNotifiationDecoder, DownlinkOperation, DownlinkOperationDecoder,
     DownlinkOperationEncoder,
 };
